fix(TablicaScreens): correct the maximum row count for fonts

+++ src/core/utils/fontmetricsutility.h
@@ -24,7 +24,7 @@ private:
     struct FontDefinition
     {
         QString family;
-        int rows;
+        float rows;
     };
 
     struct CacheKey
